ResultsResponseParser now parses null responses.
Previously, if a result in a map operation response was "null,"
the parser would quietly throw it away, and not represent it in
the ResultsResponse's results list.

For example, if you sent a touch request, Frank returned a null
result for each view that performed the request. But the
ResultsResponse returned from the parser appeared to indicate that
no views performed to the operation.

Now, if a result in the FranklyResponse is "null," the parser adds
null to the ResultsResponse's results list.